Fix build warnings, obsolete autoconf macros, and occamdoc Python 3 bugs
Build system:
- Make -fno-defer-pop, -mdynamic-no-pic, -fno-pie, -no-pie x86-only in
  m4/kroc.m4 and runtime Makefile.am files (not applicable to aarch64)
- Move -Wl,--allow-shlib-undefined to linux-only in m4/kroc.m4; remove
  from modules/cif/examples/Makefile.am (unsupported on Apple ld)
- Remove -read_only_relocs suppress from aarch64 in kroc.in/occbuild.in
- Make -mdynamic-no-pic x86-only in runtime/libkrocif/configure.ac
- Replace AC_CANONICAL_SYSTEM with AC_CANONICAL_TARGET in m4/kroc.m4
  and m4/occam-sdl.m4 (autoupdate does not touch m4 files)
- Replace AC_LANG_C/AC_LANG_SAVE/AC_LANG_RESTORE with AC_LANG_PUSH/POP
  and AC_FD_MSG with AS_MESSAGE_FD in m4/occam-opengl.m4
- Replace AC_TRY_RUN with AC_RUN_IFELSE in m4/occam-cuda.m4
- Run autoupdate on all configure.ac files; remove m4_warn(obsolete)
  blocks that autoupdate inserted (STDC_HEADERS migration notices)

occ21 warnings (legacy 1991-era code, suppress rather than rewrite):
- Add COMPILE override in tools/occ21/Makefile.am to place AM_CFLAGS
  after CFLAGS so -Wno-* flags override -Wall
- Add 18 -Wno-* suppression flags for known-harmless legacy patterns
- Fix real bugs found during review: objrdr.c pointer deref (*end),
  bind1.c format string security, vti2.c missing NULL struct fields,
  harness.c ptrdiff_t format (%td), txlib.c operator precedence,
  sort.c uintptr_t alignment check

occamdoc Python 3 fixes (applied to occamdoc.in so they survive
configure regeneration -- previous fixes were lost because only the
generated occamdoc was patched, not the template):
- parse_file: guard level < old_level with old_level is not None
  (groups push None as level; Python 3 cannot compare float < None)
- parse_comment: return None instead of "" for empty description
  (--** @module proc with no body produced "" causing false duplicate
  description error when merging .occ and .inc for same module)
- main: define encoding = "UTF-8" before it is used in XML output
- generic_tag_to_xml: use maxsplit= keyword in re.split calls

+++ tools/occ21/suplib/sort.c
@@ -33,6 +33,7 @@
 
 #include <stddef.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 #include "suplib.h"
 /* Qsort is implemented using an explicit stack rather than C recursion  */
@@ -46,7 +47,7 @@
 #define exchange(p1, p2, size)                      \
 {                                                   \
     int xsize = size;                               \
-    if (((xsize | (int) p1) & MASK)==0)             \
+    if (((xsize | (uintptr_t) p1) & MASK)==0)             \
     {   do                                          \
         {   int temp = *(int *)p1;                  \
             *(int *)p1 = *(int *)p2;                \
@@ -201,7 +202,7 @@ void sup_qsort(void *base, size_t nmemb, size_t size,
 /* The fact that I do not know how large the objects that are being      */
 /* sorted are is horrible - I do an exchange here copying things one     */
 /* word at a time or one byte at a time depending on the value of size.  */
-        if (((size | (int) b) & MASK)==0)
+        if (((size | (uintptr_t) b) & MASK)==0)
         {   int j;
             for (j=0; j<size; j+=sizeof(int))
             {   char *bx = b + j;
